ml-service/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.5.2
numpy==2.1.1
